Correct error in height measurement
It is assumed that padding and border widths can only be integral values in px units. But if CSS sets padding and border width in ems, this is likely to result in a floating value when converted to px.

Using parseInt can result in the height being overestimated by 0 to 4 px.